+++ art/src/main/java/ui/main/patterns/PatternsController.java
@@ -22,6 +22,8 @@ import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
+import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonBar;
 import javafx.scene.control.Label;
 import javafx.scene.control.SelectionMode;
 import javafx.scene.control.TableCell;
@@ -37,6 +39,7 @@ import ui.App;
 import ui.AppComponents;
 import ui.custom.controls.ButtonCell;
 import ui.custom.controls.ConditionTypeCellFactory;
+import ui.custom.controls.CustomAlert;
 import ui.custom.controls.CustomWindow;
 import ui.custom.controls.filter.FilterController;
 import ui.main.patterns.modal.PatternImportController;
@@ -272,11 +275,21 @@ public class PatternsController {
      */
     public void deleteAction() throws Exception {
         if (patternsTable.getSelectionModel().getSelectedItems() != null && patternsTable.getSelectionModel().getSelectedItems().size() != 0) {
-
-            for (AccessPattern pattern : patternsTable.getSelectionModel().getSelectedItems()) {
-                AppComponents.getDbContext().deletePattern(pattern);
+            CustomAlert customAlert;
+            if (patternsTable.getSelectionModel().getSelectedItems().size() == 1) {
+                customAlert = new CustomAlert(Alert.AlertType.CONFIRMATION, bundle.getString("deleteConfirmTitle"),
+                    bundle.getString("deleteConfirmMessage"), "Ok", "Cancel");
+            } else {
+                customAlert = new CustomAlert(Alert.AlertType.CONFIRMATION, bundle.getString("deleteMultipleConfirmTitle"),
+                    bundle.getString("deleteMultipleConfirmMessage"), "Ok", "Cancel");
             }
-            updatePatternsTable();
+
+            if (customAlert.showAndWait().get().getButtonData().equals(ButtonBar.ButtonData.OK_DONE)) {
+                for (AccessPattern pattern : patternsTable.getSelectionModel().getSelectedItems()) {
+                    AppComponents.getDbContext().deletePattern(pattern);
+                }
+                updatePatternsTable();
+            }
         }
     }
