+++ art/src/main/java/ui/main/sapsettings/SapSettingsController.java
@@ -19,6 +19,7 @@ import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.control.Alert;
+import javafx.scene.control.ButtonBar;
 import javafx.scene.control.Label;
 import javafx.scene.control.SelectionMode;
 import javafx.scene.control.TableCell;
@@ -277,12 +278,20 @@ public class SapSettingsController {
      */
     public void deleteAction() throws Exception {
         if (sapConnectionTable.getSelectionModel().getSelectedItems() != null && sapConnectionTable.getSelectionModel().getSelectedItems().size() != 0) {
-            // TODO show delete alert
-            for (SapConfiguration config : sapConnectionTable.getSelectionModel().getSelectedItems()) {
-                database.deleteSapConfig(config);
+            CustomAlert customAlert;
+            if (sapConnectionTable.getSelectionModel().getSelectedItems().size() == 1) {
+                customAlert = new CustomAlert(Alert.AlertType.CONFIRMATION, bundle.getString("deleteConfirmTitle"),
+                    bundle.getString("deleteConfirmMessage"), "Ok", "Cancel");
+            } else {
+                customAlert = new CustomAlert(Alert.AlertType.CONFIRMATION, bundle.getString("deleteMultipleConfirmTitle"),
+                    bundle.getString("deleteMultipleConfirmMessage"), "Ok", "Cancel");
             }
-            updateSapSettingsTable();
-
+            if (customAlert.showAndWait().get().getButtonData().equals(ButtonBar.ButtonData.OK_DONE)) {
+                for (SapConfiguration config : sapConnectionTable.getSelectionModel().getSelectedItems()) {
+                    database.deleteSapConfig(config);
+                }
+                updateSapSettingsTable();
+            }
         }
     }
 }
